refactor: examples to split modules
clean hooks

tweaks

fix addon not defined
