+++ src/examples/helpers.ts
@@ -1,0 +1,276 @@
+/**
+ * Helper Examples
+ * - Dialog: Dialog with data binding
+ * - Clipboard: Copy to clipboard
+ * - FilePicker: Select file from system
+ * - ProgressWindow: Show progress notification
+ * - VTable: Virtual table (see preferences)
+ */
+
+export async function dialogExample() {
+  const dialogData: { [key: string | number]: any } = {
+    inputValue: "test",
+    checkboxValue: true,
+    loadCallback: () => {
+      ztoolkit.log(dialogData, "Dialog Opened!");
+    },
+    unloadCallback: () => {
+      ztoolkit.log(dialogData, "Dialog closed!");
+    },
+  };
+  const dialogHelper = new ztoolkit.Dialog(10, 2)
+    .addCell(0, 0, {
+      tag: "h1",
+      properties: { innerHTML: "Helper Examples" },
+    })
+    .addCell(1, 0, {
+      tag: "h2",
+      properties: { innerHTML: "Dialog Data Binding" },
+    })
+    .addCell(2, 0, {
+      tag: "p",
+      properties: {
+        innerHTML:
+          "Elements with attribute 'data-bind' are binded to the prop under 'dialogData' with the same name.",
+      },
+      styles: {
+        width: "200px",
+      },
+    })
+    .addCell(3, 0, {
+      tag: "label",
+      namespace: "html",
+      attributes: {
+        for: "dialog-checkbox",
+      },
+      properties: { innerHTML: "bind:checkbox" },
+    })
+    .addCell(
+      3,
+      1,
+      {
+        tag: "input",
+        namespace: "html",
+        id: "dialog-checkbox",
+        attributes: {
+          "data-bind": "checkboxValue",
+          "data-prop": "checked",
+          type: "checkbox",
+        },
+        properties: { label: "Cell 1,0" },
+      },
+      false,
+    )
+    .addCell(4, 0, {
+      tag: "label",
+      namespace: "html",
+      attributes: {
+        for: "dialog-input",
+      },
+      properties: { innerHTML: "bind:input" },
+    })
+    .addCell(
+      4,
+      1,
+      {
+        tag: "input",
+        namespace: "html",
+        id: "dialog-input",
+        attributes: {
+          "data-bind": "inputValue",
+          "data-prop": "value",
+          type: "text",
+        },
+      },
+      false,
+    )
+    .addCell(5, 0, {
+      tag: "h2",
+      properties: { innerHTML: "Toolkit Helper Examples" },
+    })
+    .addCell(
+      6,
+      0,
+      {
+        tag: "button",
+        namespace: "html",
+        attributes: {
+          type: "button",
+        },
+        listeners: [
+          {
+            type: "click",
+            listener: (e: Event) => {
+              addon.hooks.onDialogEvents("clipboardExample");
+            },
+          },
+        ],
+        children: [
+          {
+            tag: "div",
+            styles: {
+              padding: "2.5px 15px",
+            },
+            properties: {
+              innerHTML: "example:clipboard",
+            },
+          },
+        ],
+      },
+      false,
+    )
+    .addCell(
+      7,
+      0,
+      {
+        tag: "button",
+        namespace: "html",
+        attributes: {
+          type: "button",
+        },
+        listeners: [
+          {
+            type: "click",
+            listener: (e: Event) => {
+              addon.hooks.onDialogEvents("filePickerExample");
+            },
+          },
+        ],
+        children: [
+          {
+            tag: "div",
+            styles: {
+              padding: "2.5px 15px",
+            },
+            properties: {
+              innerHTML: "example:filepicker",
+            },
+          },
+        ],
+      },
+      false,
+    )
+    .addCell(
+      8,
+      0,
+      {
+        tag: "button",
+        namespace: "html",
+        attributes: {
+          type: "button",
+        },
+        listeners: [
+          {
+            type: "click",
+            listener: (e: Event) => {
+              addon.hooks.onDialogEvents("progressWindowExample");
+            },
+          },
+        ],
+        children: [
+          {
+            tag: "div",
+            styles: {
+              padding: "2.5px 15px",
+            },
+            properties: {
+              innerHTML: "example:progressWindow",
+            },
+          },
+        ],
+      },
+      false,
+    )
+    .addCell(
+      9,
+      0,
+      {
+        tag: "button",
+        namespace: "html",
+        attributes: {
+          type: "button",
+        },
+        listeners: [
+          {
+            type: "click",
+            listener: (e: Event) => {
+              addon.hooks.onDialogEvents("vtableExample");
+            },
+          },
+        ],
+        children: [
+          {
+            tag: "div",
+            styles: {
+              padding: "2.5px 15px",
+            },
+            properties: {
+              innerHTML: "example:virtualized-table",
+            },
+          },
+        ],
+      },
+      false,
+    )
+    .addButton("Confirm", "confirm")
+    .addButton("Cancel", "cancel")
+    .addButton("Help", "help", {
+      noClose: true,
+      callback: (e) => {
+        dialogHelper.window?.alert("Help Clicked! Dialog will not be closed.");
+      },
+    })
+    .setDialogData(dialogData)
+    .open("Dialog Example");
+  addon.data.dialog = dialogHelper;
+  await dialogData.unloadLock.promise;
+  addon.data.dialog = undefined;
+  if (addon.data.alive)
+    ztoolkit.getGlobal("alert")(
+      `Close dialog with ${dialogData._lastButtonId}.\nCheckbox: ${dialogData.checkboxValue}\nInput: ${dialogData.inputValue}.`,
+    );
+  ztoolkit.log(dialogData);
+}
+
+export function clipboardExample() {
+  new ztoolkit.Clipboard()
+    .addText(
+      "![Plugin Template](https://github.com/windingwind/zotero-plugin-template)",
+      "text/unicode",
+    )
+    .addText(
+      '<a href="https://github.com/windingwind/zotero-plugin-template">Plugin Template</a>',
+      "text/html",
+    )
+    .copy();
+  ztoolkit.getGlobal("alert")("Copied!");
+}
+
+export async function filePickerExample() {
+  const path = await new ztoolkit.FilePicker(
+    "Import File",
+    "open",
+    [
+      ["PNG File(*.png)", "*.png"],
+      ["Any", "*.*"],
+    ],
+    "image.png",
+  ).open();
+  ztoolkit.getGlobal("alert")(`Selected ${path}`);
+}
+
+export function progressWindowExample() {
+  new ztoolkit.ProgressWindow(addon.data.config.addonName)
+    .createLine({
+      text: "ProgressWindow Example!",
+      type: "success",
+      progress: 100,
+    })
+    .show();
+}
+
+export function vtableExample() {
+  ztoolkit.getGlobal("alert")(
+    "See src/examples/basic.ts - registerPrefsScripts",
+  );
+}
